+++ packages/orq-rc/src/orq_ai_sdk/models/retrievedatasetop.py
@@ -79,5 +79,5 @@ class RetrieveDatasetResponseBody(BaseModel):
     created: Optional[datetime] = None
     r"""The date and time the resource was created"""
 
-    updated: Optional[datetime] = dateutil.parser.isoparse("2025-03-11T11:09:42.848Z")
+    updated: Optional[datetime] = dateutil.parser.isoparse("2025-03-12T08:05:44.240Z")
     r"""The date and time the resource was last updated"""